chore: Improve project file cleanup and avatar handling
This commit includes several improvements related to file cleanup and avatar handling in the project. The following changes were made:

In Makefile:
- Modified the 'del_uploads' target to exclude 'defaultProfileImage.jpg' from deletion when removing files in the 'uploads' directory. This ensures that the default avatar image is preserved.

In auth.service.js and auth.service.ts:
- Added the import statement for the DEFAULT_AVATAR_PATH constant.
- Set the 'avatar' field to DEFAULT_AVATAR_PATH when creating a new user. This ensures that a default avatar is assigned upon user creation.

In constants.ts:
- Defined the DEFAULT_AVATAR_PATH constant to specify the path to the default avatar image.

In users.controller.ts and users.service.ts:
- Updated avatar handling logic to prevent the deletion of the default avatar image when updating a user's avatar.
- Improved error logging and handling in the UsersController.

In UserProfileSetup.tsx:
- Refactored the code to enhance readability.
- Added comments for better code documentation.

These changes enhance the project's file cleanup process and ensure that the default avatar is correctly assigned to new users.

+++ srcs/backend/src/users/users.service.ts
@@ -14,6 +14,10 @@ import axios from 'axios';
 import { Readable } from 'stream'; // Import Readable from 'stream' module
 import * as fs from 'fs';
 import { use } from "passport";
+import { DEFAULT_AVATAR_PATH } from '../auth/constants/constants'; // Path to the default avatar image
+
+
+//defaul avatar path
 
 interface FriendRequestFromUser {
     id: number;
@@ -275,7 +279,7 @@ export class UsersService {
             const oldAvatarPath = currentUser.avatar;
 
             // If there is an old avatar, attempt to delete it
-            if (oldAvatarPath) {
+            if (oldAvatarPath && oldAvatarPath !== DEFAULT_AVATAR_PATH) {
                 try {
                     // Use fs.unlinkSync for synchronous file deletion
                     fs.unlinkSync(oldAvatarPath);
